+++ mycodo/inputs/winsen_zh03b.py
@@ -114,7 +114,6 @@ class InputModule(AbstractInput):
                         self.fan_state = self.DormantMode('run')
 
                     time.sleep(0.1)
-                    self.get_measurement(silent=True)  # Throw out first measurement
 
                 except serial.SerialException:
                     self.logger.exception('Opening serial')
@@ -172,7 +171,7 @@ class InputModule(AbstractInput):
             self.read()
         return self._pm_10_0
 
-    def get_measurement(self, silent=False):
+    def get_measurement(self):
         """ Gets the WINSEN_ZH03B's Particulate concentration in μg/m^3 via UART """
         if not self.serial_device:  # Don't measure if device isn't validated
             return None, None, None
@@ -204,16 +203,13 @@ class InputModule(AbstractInput):
 
             if pm_1_0 > 1000:
                 pm_1_0 = None
-                if not silent:
-                    self.logger.error("PM1 measurement out of range (over 1000 ug/m^3)")
+                self.logger.error("PM1 measurement out of range (over 1000 ug/m^3)")
             if pm_2_5 > 1000:
                 pm_2_5 = None
-                if not silent:
-                    self.logger.error("PM2.5 measurement out of range (over 1000 ug/m^3)")
+                self.logger.error("PM2.5 measurement out of range (over 1000 ug/m^3)")
             if pm_10_0 > 1000:
                 pm_10_0 = None
-                if not silent:
-                    self.logger.error("PM10 measurement out of range (over 1000 ug/m^3)")
+                self.logger.error("PM10 measurement out of range (over 1000 ug/m^3)")
 
         except:
             self.logger.exception("Exception while reading")